feat: Apply default values when using package API
chore(refactor): Move config related things unrelated
to CLI specifics into config.js module. Make it easier
to spot differences between external user config schema
and runtime config schema.
