+++ lib/api.js
@@ -1,0 +1,1 @@
+export  {runViaApi as run, getSlugger} from "./main.js";
