+++ lib/main.js
@@ -2,10 +2,8 @@ import GitHubSlugger from "github-slugger";
 import { exportGlossaries } from "./exporter.js";
 import { importGlossaries } from "./importer.js";
 import { newContext } from "./model/context.js";
-import {
-    readDocumentFiles,
-    readGlossaries
-} from "./reader.js";
+import { validateConfig, getRunnableConfig } from "./model/config.js";
+import { readDocumentFiles, readGlossaries } from "./reader.js";
 import {
     copyBaseDirToOutDir,
     writeDocumentFiles,
@@ -15,8 +13,34 @@ import {
     writeTestOutput
 } from "./writer.js";
 
-export async function run(glossarifyMdConf) {
-    let context = await newContext(glossarifyMdConf);
+
+/**
+ * @param {object} confUser User-provided configuration adhering to the external configuration schema.
+ * @returns {Promise} a promise from an async execution resolving to an object representing the execution context.
+ */
+export async function runViaApi(confUser) {
+    const conf = getRunnableConfig(confUser);
+    conf.force = false;
+    return run(conf);
+}
+
+
+/**
+ * @param {object} confUser User-provided configuration adhering to the external configuration schema. Assumed to
+ * have been derived from applying config.js#getRunConfig() already.
+ * @param {object} forceFlag A flag indicating presence of the --force CLI argument.
+ * @returns {Promise} a promise from an async execution resolving to an object representing the execution context.
+ */
+export async function runViaCli(conf, forceFlag) {
+    conf.force = forceFlag;
+    return run(conf);
+}
+
+async function run(conf) {
+
+    validateConfig(conf);
+
+    const context = await newContext(conf);
     await copyBaseDirToOutDir(context);
     await importGlossaries(context);
     await readGlossaries(context);
