+++ lib/model/context.js
@@ -4,7 +4,7 @@ import { Glob } from "glob";
 import { relativeFromTo, toForwardSlash } from "../path/tools.js";
 import { init as initCollator } from "../text/collator.js";
 import { Glossary } from "./glossary.js";
-
+import { getRuntimeConfig } from "./config.js";
 
 /**
  *
@@ -31,48 +31,7 @@ class Context {
          * @type {VFile}
          */
         this.writeFiles = [];
-        const conf_ = this.conf = {
-            ...conf
-            ,baseDir: toForwardSlash(conf.baseDir || "")
-            ,outDir: toForwardSlash(conf.outDir  || "")
-            ,indexing: {
-                ...conf.indexing
-                // Excluding certain headingDepths in (cross-)linking
-                ,headingDepths: arrayToMap(conf.indexing.headingDepths)
-            }
-            ,linking: {
-                ...conf.linking
-                ,headingDepths: arrayToMap(conf.linking.headingDepths)
-                ,limitByTermOrigin: arrayToMap(conf.linking.limitByTermOrigin)
-                ,pathRewrites: reverseMultiValueMap(conf.linking.pathRewrites)
-                ,sortAlternatives: {
-                    by: "glossary-filename"
-                    ,...conf.linking.sortAlternatives
-                }
-            }
-        };
-
-        // limit link creation for alternative definitions
-        const altLinks = conf.linking.limitByAlternatives;
-        if (Math.abs(altLinks) > 95) {
-            conf_.linking.limitByAlternatives = Math.sign(altLinks) * 95;
-        }
-        const sortAlternatives = conf_.linking.sortAlternatives;
-        if (sortAlternatives.by === "glossary-ref-count") {
-            conf_.linking.sortAlternatives = { perSectionDepth: 2, ...sortAlternatives };
-        }
-        if (conf.generateFiles.listOfFigures) {
-            conf_.generateFiles.listOfFigures = { class: "figure", title: "Figures", ...conf.generateFiles.listOfFigures };
-            conf_.generateFiles.listOf.push(conf.generateFiles.listOfFigures);
-        }
-        if (conf.generateFiles.listOfTables) {
-            conf_.generateFiles.listOfTables = { class: "table", title: "Tables", ...conf.generateFiles.listOfTables };
-            conf_.generateFiles.listOf.push(conf.generateFiles.listOfTables);
-        }
-        if (conf_.unified.rcPath) {
-            conf_.unified.rcPath = toForwardSlash(path.resolve(conf.baseDir, conf.unified.rcPath));
-        }
-
+        this.conf = getRuntimeConfig(conf);
     }
 
     resolvePath(relativePath) {
@@ -85,46 +44,6 @@ class Context {
     }
 }
 
-function arrayToMap(array) {
-    return array.reduce((prev, curr) => {
-        prev[curr] = true;
-        return prev;
-    }, {});
-}
-
-/**
- * Transforms a map
- * {
- *    "key1": ["value1", "value2", "value3"],
- *    "key2": ["value1", "valueA", "valueB"],
- *    "key3": "value%"
- * }
- * to a map
- * {
- *    "value1": "key2",
- *    "value2": "key",
- *    "value3": "key",
- *    "valueA": "key2",
- *    "valueB": "key2",
- *    "value%": "key3"
- * }
- */
-function reverseMultiValueMap(input) {
-    const output = {};
-    for (const key in input) {
-        if (input[key]) {
-            const values = [].concat(input[key]); // [1]
-            for (let i = 0, len = values.length; i < len; i++) {
-                const value = values[i];
-                output[value] = key;
-            }
-        }
-    }
-    return output;
-
-    // ___/ Implementation Notes \___
-    // [1] use .concat() to wrap %value% into [ "%value%" ] prior to forEach()
-}
 
 async function unglobGlossaries(context) {
     const { baseDir, glossaries, excludeFiles } = context.conf;
